coloco tr y td de table
